requirements.txt
Flask == 1.0.2
# Flask-RESTful == 0.3.6
# Flask-Cors == 3.0.7

Flask-Migrate == 2.3.1
Flask-Script == 2.0.6
SQLAlchemy == 1.2.16
flask-sqlalchemy == 2.3.2
marshmallow == 2.16.3
flask-marshmallow == 0.9.0
marshmallow-sqlalchemy == 0.15.0
# psycopg2 == 2.8.6
psycopg2-binary
Flask-Mail == 0.9.1
flask-bcrypt == 0.7.1
Flask-Uploads == 0.2.1
simplejson == 3.16.0
requests == 2.26.0
phonenumbers == 8.10.4
Flask-Principal == 0.4.0
Flask-Security == 3.0.0
schedule == 0.6.0
html2text == 2020.1.16
num2words == 0.5.10
pdfkit == 0.6.1
PyVirtualDisplay == 1.3.2
ics==0.7.2
flask_jwt_extended == 4.0.0
python-dotenv==0.20.0
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Jinja2==2.11.2
MarkupSafe==1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1